Daybook finish without testing

+++ src/modules/daybook/views/EntryView.vue
@@ -8,12 +8,20 @@
             </div>
 
             <div>
+
+                <input type="file"
+                    @change="onSelectedImage"
+                    ref="imageSelector"
+                    v-show="false"
+                    accept="image/png image/jpeg"
+                >
+
                 <button v-if="entry.id" class="btn btn-danger mx-2" @click="onDeleteEntry">
                     Borrar
                     <i class="fa fa-trash-alt"></i>
                 </button>
 
-                <button class="btn btn-primary">
+                <button class="btn btn-primary" @click="onSelectImage">
                     Subir foto
                     <i class="fa fa-upload"></i>
                 </button>
@@ -24,7 +32,15 @@
             <textarea v-model="entry.text" placeholder="¿Qué sucedió hoy?"></textarea>
         </div>
 
-        <img src="https://cdn.forbes.com.mx/2022/01/Cabo-San-lucas-640x360.jpg" alt="entry-picture"
+        <img 
+            v-if="entry.picture && !localImage"
+            :src="entry.picture" 
+            alt="entry-picture"
+            class="img-thumbnail">
+        <img 
+            v-if="localImage"
+            :src="localImage"
+            alt="entry-picture"
             class="img-thumbnail">
 
     </template>
@@ -36,7 +52,9 @@
 <script>
 import { defineAsyncComponent } from 'vue';
 import { mapGetters, mapActions } from 'vuex';
-import getDayMonthYear from "../helpers/getDayMonthYear";
+import Swal from 'sweetalert2'
+import getDayMonthYear from '../helpers/getDayMonthYear';
+import uploadImage from '../helpers/uploadImage';
 
 export default {
     props: {
@@ -50,7 +68,9 @@ export default {
     },
     data() {
         return {
-            entry: null
+            entry: null,
+            localImage: null,
+            file: null
         }
     },
 
@@ -88,24 +108,66 @@ export default {
             this.entry = entry
         },
         async saveEntry() {
+            new Swal({
+                title: "Espere por favor.",
+                allowOutsideClick: false
+            })
+
+            Swal.showLoading()
+
+            const picture = await uploadImage( this.file )
+
+            this.entry.picture = picture
+
             if (this.entry.id) {
                 // Actualizar 
                 await this.updateEntry(this.entry)
             } else {
                 // Crear una nueva entrada
-                console.log('Post de una nueva entrada')
                 const id = await this.createEntry(this.entry)
                 this.$router.push({ name: 'entry', params: { id } })
-
             }
+            this.file = null
+            Swal.fire("Guardado", "Entrada registrada con éxito", "success")
         },
         async onDeleteEntry() {
-            console.log("delete", this.entry)
-            await this.deleteEntry(this.entry.id)
-            // redireccionar al usuario fuera de aqui
-            this.$router.push({ name: 'no-entry' })
+            
+            const result = await  Swal.fire({
+                title: "¿Estás seguro?",
+                text: "Una vezk borrado, no se puede recuperar.",
+                showDenyButton: true,
+                confirmButtonText: "Sí, estoy seguro"
+            })
 
+            if( result.isConfirmed ){
+                new Swal({
+                    title: "Espere por favor",
+                    allowOutsideClick: false
+                })
+                Swal.showLoading()
+                await this.deleteEntry(this.entry.id)
+                this.$router.push({ name: 'no-entry' })
+
+                Swal.fire("Eliminado", "", "success")
+            }
         },
+
+        onSelectedImage($event){
+            const file = $event.target.files[0]
+            if( !file ) {
+                this.localImage = null
+                this.file = null
+                return 
+            }
+
+            this.file = file
+            const fr = new FileReader()
+            fr.onload = () => this.localImage = fr.result
+            fr.readAsDataURL( file )
+        },
+        onSelectImage(){
+            this.$refs.imageSelector.click()
+        }
     },
 
     created() {
